Datagen grist cost builders cleanup and moved to api
