+++ src/main/java/com/mraof/minestuck/api/alchemy/recipe/ContainerGristCostBuilder.java
@@ -1,9 +1,10 @@
-package com.mraof.minestuck.data.recipe;
+package com.mraof.minestuck.api.alchemy.recipe;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.gson.JsonObject;
 import com.mojang.serialization.JsonOps;
 import com.mraof.minestuck.api.alchemy.ImmutableGristSet;
+import com.mraof.minestuck.data.recipe.AdvancementFreeRecipe;
 import com.mraof.minestuck.item.crafting.MSRecipeTypes;
 import com.mraof.minestuck.api.alchemy.GristType;
 import com.mraof.minestuck.api.alchemy.DefaultImmutableGristSet;
@@ -20,59 +21,61 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.Objects;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
+/**
+ * Used to datagen a grist cost where the cost of the container item is automatically added to the recipe cost. Typically used for buckets.
+ */
+@ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-public class GristCostRecipeBuilder
+public final class ContainerGristCostBuilder
 {
 	private static final Logger LOGGER = LogManager.getLogger();
 	
+	@Nullable
 	private final ResourceLocation defaultName;
 	private final Ingredient ingredient;
 	private final ImmutableMap.Builder<GristType, Long> costBuilder = ImmutableMap.builder();
+	@Nullable
 	private Integer priority = null;
 	
-	public static GristCostRecipeBuilder of(TagKey<Item> tag)
+	public static ContainerGristCostBuilder of(TagKey<Item> tag)
 	{
 		ResourceLocation tagId = tag.location();
-		return new GristCostRecipeBuilder(new ResourceLocation(tagId.getNamespace(), tagId.getPath()+"_tag"), Ingredient.of(tag));
+		return new ContainerGristCostBuilder(new ResourceLocation(tagId.getNamespace(), tagId.getPath()+"_tag"), Ingredient.of(tag));
 	}
 	
-	public static GristCostRecipeBuilder of(ItemLike item)
+	public static ContainerGristCostBuilder of(ItemLike item)
 	{
-		return new GristCostRecipeBuilder(ForgeRegistries.ITEMS.getKey(item.asItem()), Ingredient.of(item));
+		return new ContainerGristCostBuilder(ForgeRegistries.ITEMS.getKey(item.asItem()), Ingredient.of(item));
 	}
 	
-	public static GristCostRecipeBuilder of(Ingredient ingredient)
+	public static ContainerGristCostBuilder of(Ingredient ingredient)
 	{
-		return new GristCostRecipeBuilder(ingredient);
+		return new ContainerGristCostBuilder(null, ingredient);
 	}
 	
-	protected GristCostRecipeBuilder(Ingredient ingredient)
-	{
-		this(null, ingredient);
-	}
-	
-	protected GristCostRecipeBuilder(ResourceLocation defaultName, Ingredient ingredient)
+	private ContainerGristCostBuilder(@Nullable ResourceLocation defaultName, Ingredient ingredient)
 	{
 		this.defaultName = defaultName;
 		this.ingredient = ingredient;
 	}
 	
-	public GristCostRecipeBuilder grist(Supplier<GristType> type, long amount)
+	public ContainerGristCostBuilder grist(Supplier<GristType> type, long amount)
 	{
 		return grist(type.get(), amount);
 	}
 	
-	public GristCostRecipeBuilder grist(GristType type, long amount)
+	public ContainerGristCostBuilder grist(GristType type, long amount)
 	{
 		costBuilder.put(type, amount);
 		return this;
 	}
 	
-	public GristCostRecipeBuilder priority(int priority)
+	public ContainerGristCostBuilder priority(int priority)
 	{
 		this.priority = priority;
 		return this;
@@ -95,21 +98,8 @@ public class GristCostRecipeBuilder
 		recipeSaver.accept(new Result(new ResourceLocation(id.getNamespace(), "grist_costs/"+id.getPath()), ingredient, new DefaultImmutableGristSet(costBuilder), priority));
 	}
 	
-	public static class Result implements FinishedRecipe
+	private record Result(ResourceLocation id, Ingredient ingredient, ImmutableGristSet cost, @Nullable Integer priority) implements AdvancementFreeRecipe
 	{
-		public final ResourceLocation id;
-		public final Ingredient ingredient;
-		public final ImmutableGristSet cost;
-		public final Integer priority;
-		
-		public Result(ResourceLocation id, Ingredient ingredient, ImmutableGristSet cost, Integer priority)
-		{
-			this.id = id;
-			this.ingredient = ingredient;
-			this.cost = cost;
-			this.priority = priority;
-		}
-		
 		@Override
 		public void serializeRecipeData(JsonObject jsonObject)
 		{
@@ -128,21 +118,7 @@ public class GristCostRecipeBuilder
 		@Override
 		public RecipeSerializer<?> getType()
 		{
-			return MSRecipeTypes.GRIST_COST.get();
-		}
-		
-		@Nullable
-		@Override
-		public JsonObject serializeAdvancement()
-		{
-			return null;
-		}
-		
-		@Nullable
-		@Override
-		public ResourceLocation getAdvancementId()
-		{
-			return new ResourceLocation("");
+			return MSRecipeTypes.CONTAINER_GRIST_COST.get();
 		}
 	}
 }
