+++ src/main/java/com/mraof/minestuck/api/alchemy/recipe/SourceGristCostBuilder.java
@@ -1,13 +1,16 @@
-package com.mraof.minestuck.data.recipe;
+package com.mraof.minestuck.api.alchemy.recipe;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
+import com.mojang.serialization.JsonOps;
 import com.mraof.minestuck.api.alchemy.DefaultImmutableGristSet;
 import com.mraof.minestuck.api.alchemy.GristType;
 import com.mraof.minestuck.api.alchemy.ImmutableGristSet;
+import com.mraof.minestuck.data.recipe.AdvancementFreeRecipe;
 import com.mraof.minestuck.item.crafting.MSRecipeTypes;
 import com.mraof.minestuck.alchemy.recipe.generator.SourceGristCost;
+import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.data.recipes.FinishedRecipe;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
@@ -16,20 +19,33 @@ import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.minecraft.world.level.ItemLike;
 import net.minecraftforge.registries.ForgeRegistries;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
+import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
-public class SourceGristCostBuilder
+/**
+ * Used to datagen a grist cost that sources the cost from different items with some modification.
+ */
+@ParametersAreNonnullByDefault
+@MethodsReturnNonnullByDefault
+public final class SourceGristCostBuilder
 {
+	private static final Logger LOGGER = LogManager.getLogger();
+	
+	@Nullable
 	private final ResourceLocation defaultName;
 	private final Ingredient ingredient;
 	private final ImmutableMap.Builder<GristType, Long> costBuilder = ImmutableMap.builder();
 	private final List<String> sources = new ArrayList<>();
 	private float multiplier = 1;
+	@Nullable
 	private Integer priority = null;
 	
 	public static SourceGristCostBuilder of(TagKey<Item> tag)
@@ -45,15 +61,10 @@ public class SourceGristCostBuilder
 	
 	public static SourceGristCostBuilder of(Ingredient ingredient)
 	{
-		return new SourceGristCostBuilder(ingredient);
+		return new SourceGristCostBuilder(null, ingredient);
 	}
 	
-	protected SourceGristCostBuilder(Ingredient ingredient)
-	{
-		this(null, ingredient);
-	}
-	
-	protected SourceGristCostBuilder(ResourceLocation defaultName, Ingredient ingredient)
+	private SourceGristCostBuilder(@Nullable ResourceLocation defaultName, Ingredient ingredient)
 	{
 		this.defaultName = defaultName;
 		this.ingredient = ingredient;
@@ -109,32 +120,35 @@ public class SourceGristCostBuilder
 	
 	public void build(Consumer<FinishedRecipe> recipeSaver, ResourceLocation id)
 	{
-		recipeSaver.accept(new Result(new ResourceLocation(id.getNamespace(), "grist_costs/"+id.getPath()), ingredient, sources, multiplier, new DefaultImmutableGristSet(costBuilder), priority));
+		recipeSaver.accept(new Result(id.withPrefix("grist_costs/"), ingredient, priority, sources, multiplier, new DefaultImmutableGristSet(costBuilder)));
 	}
 	
-	public static class Result extends GristCostRecipeBuilder.Result
+	private record Result(ResourceLocation id, Ingredient ingredient, @Nullable Integer priority,
+						  List<String> sources, float multiplier, ImmutableGristSet cost) implements AdvancementFreeRecipe
 	{
-		private final List<String> sources;
-		private final float multiplier;
-		
-		public Result(ResourceLocation id, Ingredient ingredient, List<String> sources, float multiplier, ImmutableGristSet cost, Integer priority)
-		{
-			super(id, ingredient, cost, priority);
-			this.sources = sources;
-			this.multiplier = multiplier;
-		}
-		
 		@Override
 		public void serializeRecipeData(JsonObject jsonObject)
 		{
-			super.serializeRecipeData(jsonObject);
+			jsonObject.add("ingredient", ingredient.toJson());
+			
+			if(priority != null)
+				jsonObject.addProperty("priority", priority);
+			
 			JsonArray sourceArray = new JsonArray();
 			sources.forEach(sourceArray::add);
 			jsonObject.add("sources", sourceArray);
+			
 			if(multiplier != 1)
 				jsonObject.addProperty("multiplier", multiplier);
+			
+			jsonObject.add("grist_cost", ImmutableGristSet.MAP_CODEC.encodeStart(JsonOps.INSTANCE, cost).getOrThrow(false, LOGGER::error));
 		}
 		
+		@Override
+		public ResourceLocation getId()
+		{
+			return id;
+		}
 		@Override
 		public RecipeSerializer<?> getType()
 		{
