+++ src/main/java/com/mraof/minestuck/api/alchemy/recipe/WildcardGristCostBuilder.java
@@ -1,7 +1,9 @@
-package com.mraof.minestuck.data.recipe;
+package com.mraof.minestuck.api.alchemy.recipe;
 
 import com.google.gson.JsonObject;
+import com.mraof.minestuck.data.recipe.AdvancementFreeRecipe;
 import com.mraof.minestuck.item.crafting.MSRecipeTypes;
+import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.data.recipes.FinishedRecipe;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
@@ -12,43 +14,54 @@ import net.minecraft.world.level.ItemLike;
 import net.minecraftforge.registries.ForgeRegistries;
 
 import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.Objects;
 import java.util.function.Consumer;
 
-public class UnavailableGristCostBuilder
+/**
+ * Used to datagen a grist cost which accepts an amount of any grist type chosen by the player.
+ */
+@ParametersAreNonnullByDefault
+@MethodsReturnNonnullByDefault
+@SuppressWarnings("unused")
+public final class WildcardGristCostBuilder
 {
+	@Nullable
 	private final ResourceLocation defaultName;
 	private final Ingredient ingredient;
+	private long cost = 0;
+	@Nullable
 	private Integer priority = null;
 	
-	public static UnavailableGristCostBuilder of(TagKey<Item> tag)
+	public static WildcardGristCostBuilder of(TagKey<Item> tag)
 	{
 		ResourceLocation tagId = tag.location();
-		return new UnavailableGristCostBuilder(new ResourceLocation(tagId.getNamespace(), tagId.getPath()+"_tag"), Ingredient.of(tag));
+		return new WildcardGristCostBuilder(new ResourceLocation(tagId.getNamespace(), tagId.getPath()+"_tag"), Ingredient.of(tag));
 	}
 	
-	public static UnavailableGristCostBuilder of(ItemLike item)
+	public static WildcardGristCostBuilder of(ItemLike item)
 	{
-		return new UnavailableGristCostBuilder(ForgeRegistries.ITEMS.getKey(item.asItem()), Ingredient.of(item));
+		return new WildcardGristCostBuilder(ForgeRegistries.ITEMS.getKey(item.asItem()), Ingredient.of(item));
 	}
 	
-	public static UnavailableGristCostBuilder of(Ingredient ingredient)
+	public static WildcardGristCostBuilder of(Ingredient ingredient)
 	{
-		return new UnavailableGristCostBuilder(ingredient);
+		return new WildcardGristCostBuilder(null, ingredient);
 	}
 	
-	protected UnavailableGristCostBuilder(Ingredient ingredient)
-	{
-		this(null, ingredient);
-	}
-	
-	protected UnavailableGristCostBuilder(ResourceLocation defaultName, Ingredient ingredient)
+	private WildcardGristCostBuilder(@Nullable ResourceLocation defaultName, Ingredient ingredient)
 	{
 		this.defaultName = defaultName;
 		this.ingredient = ingredient;
 	}
 	
-	public UnavailableGristCostBuilder priority(int priority)
+	public WildcardGristCostBuilder cost(long wildcardCost)
+	{
+		cost = wildcardCost;
+		return this;
+	}
+	
+	public WildcardGristCostBuilder priority(int priority)
 	{
 		this.priority = priority;
 		return this;
@@ -68,26 +81,18 @@ public class UnavailableGristCostBuilder
 	
 	public void build(Consumer<FinishedRecipe> recipeSaver, ResourceLocation id)
 	{
-		recipeSaver.accept(new Result(new ResourceLocation(id.getNamespace(), "grist_costs/"+id.getPath()), ingredient, priority));
+		if(this.cost == 0)
+			throw new IllegalStateException("Must set the wildcard cost before building!");
+		recipeSaver.accept(new Result(new ResourceLocation(id.getNamespace(), "grist_costs/"+id.getPath()), ingredient, cost, priority));
 	}
 	
-	public static class Result implements FinishedRecipe
+	private record Result(ResourceLocation id, Ingredient ingredient, long cost, @Nullable Integer priority) implements AdvancementFreeRecipe
 	{
-		public final ResourceLocation id;
-		public final Ingredient ingredient;
-		public final Integer priority;
-		
-		public Result(ResourceLocation id, Ingredient ingredient, Integer priority)
-		{
-			this.id = id;
-			this.ingredient = ingredient;
-			this.priority = priority;
-		}
-		
 		@Override
 		public void serializeRecipeData(JsonObject jsonObject)
 		{
 			jsonObject.add("ingredient", ingredient.toJson());
+			jsonObject.addProperty("grist_cost", cost);
 			if(priority != null)
 				jsonObject.addProperty("priority", priority);
 		}
@@ -101,21 +106,7 @@ public class UnavailableGristCostBuilder
 		@Override
 		public RecipeSerializer<?> getType()
 		{
-			return MSRecipeTypes.UNAVAILABLE_GRIST_COST.get();
-		}
-		
-		@Nullable
-		@Override
-		public JsonObject serializeAdvancement()
-		{
-			return null;
-		}
-		
-		@Nullable
-		@Override
-		public ResourceLocation getAdvancementId()
-		{
-			return new ResourceLocation("");
+			return MSRecipeTypes.WILDCARD_GRIST_COST.get();
 		}
 	}
 }
